fix 10

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -4,7 +4,7 @@ DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS FLOAT
 DETERMINISTIC
 BEGIN
-    DECLARE result INT;
+    DECLARE result FLOAT;
     IF b = 0 THEN
         SET result = 0;
     ELSE
